feat: permet l'ajout de parcelle à l'étape draft pour les agris
