+++ src/pages/certification/exploitations/[id]/index.vue
@@ -17,7 +17,7 @@ meta:
 
         <OperatorSummary :operator="record.operator" :record="recordStore.record" />
         <CertificationSummary v-if="!needsSetup" :operator="record.operator" :record="recordStore.record" :features="featureStore.collection" :validation-rules="validationRules" />
-        <FeaturesTable v-if="!needsSetup" :operator="record.operator" :features="featureStore.collection" :edit-form="EditForm" :validation-rules="validationRules" :mass-actions="massActions" add-parcel-route-name="certification-exploitations-id-ajout-parcelle" />
+        <FeaturesTable v-if="!needsSetup" :operator="record.operator" :features="featureStore.collection" :edit-form="EditForm" :validation-rules="validationRules" :mass-actions="massActions" />
         <OperatorSetup v-if="needsSetup" :operator="record.operator" />
       </div>
 
@@ -59,7 +59,6 @@ meta:
 
 <script setup>
 import { computed, markRaw, readonly, ref, shallowRef } from 'vue'
-import { getOperatorParcelles } from '@/cartobio-api'
 import bbox from '@turf/bbox'
 
 import MapContainer from '@/components/Map/MapContainer.vue'
@@ -81,7 +80,7 @@ import surroundingsStyle from "@/map-styles/surroundings.json";
 import cadastreStyle from "@/map-styles/cadastre.json";
 import SurroundingsLegend from "@/components/Map/SurroundingsLegend.vue";
 
-const props = defineProps({
+defineProps({
   id: {
     type: String,
     required: true
@@ -90,6 +89,7 @@ const props = defineProps({
 
 const featureStore = useFeaturesStore()
 const recordStore = useRecordStore()
+const record = recordStore.record
 const mapRef = shallowRef(null);
 const showSurroundings = ref(false);
 const isMapFullSize = ref(false)
@@ -103,11 +103,6 @@ const massActions = [
   { label: 'Ajouter une date d\'engagement', component: markRaw(EngagementDateModal) },
   { label: 'Modifier le niveau de conversion', component: markRaw(EngagementLevelModal) },
 ]
-
-const record = await getOperatorParcelles(props.id)
-
-recordStore.update(record)
-featureStore.setAll(record.parcelles.features)
 
 const needsSetup = computed(() => featureStore.collection.features.length === 0 && !recordStore.record.metadata.source)
 const mapBounds = computed(() => needsSetup.value ? [] : bbox(featureStore.collection))
