+++ src/pages/exploitations/[id]/ajout-parcelle.vue
@@ -1,0 +1,86 @@
+<route lang="yaml">
+  meta:
+    requiredPermissions: ['canAddParcelle']
+    seo:
+      title: Ajouter une parcelle
+</route>
+
+<template>
+  <section class="fr-container fr-py-6w">
+    <div class="fr-grid-row fr-grid-row--gutters">
+      <div class="fr-col-6">
+        <p>
+          <router-link :to="backLink" class="fr-btn fr-btn--tertiary-no-outline fr-btn--icon-left fr-icon-arrow-left-line">
+            Retour à l'exploitation
+          </router-link>
+        </p>
+
+        <OperatorSummary :operator="record.operator" :record="recordStore.record" disable-actions />
+        <FeatureAddFlow :operator="record.operator" :record="recordStore.record" :collection="featureStore.collection" @update="pendingChangesCollection = $event" />
+      </div>
+
+      <div class="fr-col-6">
+        <MapContainer v-if="!needsSetup" :options="{ interactive: true }" class="map" :bounds="mapBounds" @load="setupWatchers">
+          <GeojsonLayer :style="baseStyle" name="base" />
+          <GeojsonLayer :style="cadastreStyle" name="cadastre" />
+          <GeojsonLayer :data="featureStore.collection" name="parcellaire-operateur" />
+          <GeojsonLayer :data="pendingChangesCollection" name="pending-changes" :fill="{ 'fill-color': '#FDE2B5', 'fill-outline-color': '#000091', 'fill-opacity': 0.8 }" />
+        </MapContainer>
+        <SetupOverlay v-else>
+          Parcellaire inconnu
+        </SetupOverlay>
+      </div>
+    </div>
+  </section>
+</template>
+
+<script setup>
+import { computed, shallowRef } from 'vue'
+import bbox from '@turf/bbox'
+
+import MapContainer from '@/components/Map/MapContainer.vue'
+import GeojsonLayer from '@/components/Map/GeojsonLayer.vue'
+import SetupOverlay from '@/components/Map/SetupOverlay.vue'
+import OperatorSummary from '@/components/Operator/Summary.vue'
+import FeatureAddFlow from '@/components/Features/AddFlow/index.vue'
+
+import baseStyle from '@/map-styles/base.json'
+import cadastreStyle from '@/map-styles/cadastre.json'
+
+import { useFeaturesStore, useRecordStore, useUserStore } from '@/stores/index.js'
+import { ROLES } from "@/stores/user.js"
+
+const props = defineProps({
+  id: {
+    type: String,
+    required: true
+  }
+})
+
+const featureStore = useFeaturesStore()
+const recordStore = useRecordStore()
+const userStore = useUserStore()
+const record = recordStore.record
+
+const pendingChangesCollection = shallowRef()
+const needsSetup = computed(() => featureStore.collection.features.length === 0 && !record.metadata.source)
+const mapBounds = computed(() => needsSetup.value ? [] : bbox(featureStore.collection))
+
+const backLink = computed(() => userStore.role === ROLES.OC ?
+    `/certification/exploitations/${props.id}` : '/exploitation/parcellaire')
+
+function setupWatchers (map) {
+  featureStore.bindMaplibreFeatureState({ map, source: 'parcellaire-operateur' })
+  featureStore.bindMaplibreInteractions({ map, layer: 'parcellaire-operateur-geometry' })
+}
+</script>
+
+<style>
+.map {
+  background: #ccc;
+  height: min(100vh, 600px);
+  position: sticky;
+  top: 0;
+  width: 100%;
+}
+</style>
